requirements.txt
pillow
imageio
openai
